Rename file one_d_func.py to simple_sim.py
