+++ libensemble/tests/functionality_tests/test_1d_sampling_no_comms_given.py
@@ -20,7 +20,7 @@ from libensemble import Ensemble
 from libensemble.gen_funcs.sampling import latin_hypercube_sample as gen_f
 
 # Import libEnsemble items for this test
-from libensemble.sim_funcs.one_d_func import one_d_example as sim_f
+from libensemble.sim_funcs.simple_sim import one_d_example as sim_f
 from libensemble.specs import ExitCriteria, GenSpecs, LibeSpecs, SimSpecs
 from libensemble.tools import check_npy_file_exists
