+++ libensemble/tests/functionality_tests/test_1d_splitcomm.py
@@ -19,7 +19,7 @@ from libensemble.gen_funcs.sampling import latin_hypercube_sample as gen_f
 
 # Import libEnsemble items for this test
 from libensemble.libE import libE
-from libensemble.sim_funcs.one_d_func import one_d_example as sim_f
+from libensemble.sim_funcs.simple_sim import one_d_example as sim_f
 from libensemble.tests.regression_tests.common import mpi_comm_split
 from libensemble.tools import add_unique_random_streams, parse_args, save_libE_output
