+++ libensemble/utils/validators.py
@@ -171,7 +171,7 @@ if pydanticV1:
     @root_validator
     def simf_set_in_out_from_attrs(cls, values):
         if not values.get("sim_f"):
-            from libensemble.sim_funcs.one_d_func import one_d_example
+            from libensemble.sim_funcs.simple_sim import one_d_example
 
             values["sim_f"] = one_d_example
         if hasattr(values.get("sim_f"), "inputs") and not values.get("inputs"):
